Import gebarateList function,saveTo LocalStorage
